+++ app/Http/Controllers/PostsController.php
@@ -14,15 +14,4 @@ class PostsController extends Controller
             'posts' => $posts,
         ]);
     }
-
-    public function show($id)
-    {
-        $user = User::findOrFail($id);
-        $posts = $user->posts()->orderBy('id', 'desc')->paginate(10);
-        $data=[
-            'user' => $user,
-            'posts' => $posts,
-        ];
-        return view('users.show',$data);
-    }
 }
