+++ resources/views/commons/header.blade.php
@@ -8,7 +8,7 @@
             <ul class="navbar-nav mr-auto"></ul>
             <ul class="navbar-nav">
                 @if (Auth::check())
-                <li class="nav-item"><a href="{{ route('posts.show', Auth::id()) }}" class="nav-link text-light">{{ Auth::user()->name }}</a></li>
+                <li class="nav-item"><a href="{{ route('users.show', Auth::id()) }}" class="nav-link text-light">{{ Auth::user()->name }}</a></li>
                 <li class="nav-item"><a href="" class="nav-link text-light">ログアウト</a></li>
                 @else
                 <li class="nav-item"><a href="" class="nav-link text-light">ログイン</a></li>
